Rename Partnerships settings classes to Partner Benefits
