+++ subscriptions/subscriptions-impl/src/main/java/com/duckduckgo/subscriptions/impl/settings/views/PartnerBenefitSettingViewModel.kt
@@ -29,7 +29,7 @@ import com.duckduckgo.subscriptions.impl.SubscriptionsFeature
 import com.duckduckgo.subscriptions.impl.internal.PartnershipsHubUrlProvider
 import com.duckduckgo.subscriptions.impl.pixels.SubscriptionPixelSender
 import com.duckduckgo.subscriptions.impl.repository.isActive
-import com.duckduckgo.subscriptions.impl.settings.views.PartnershipsSettingViewModel.Command.OpenPartnershipsHub
+import com.duckduckgo.subscriptions.impl.settings.views.PartnerBenefitSettingViewModel.Command.OpenPartnershipsHub
 import kotlinx.coroutines.channels.BufferOverflow
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
@@ -46,7 +46,7 @@ import javax.inject.Inject
 
 @SuppressLint("NoLifecycleObserver")
 @ContributesViewModel(ViewScope::class)
-class PartnershipsSettingViewModel @Inject constructor(
+class PartnerBenefitSettingViewModel @Inject constructor(
     private val subscriptions: Subscriptions,
     private val subscriptionsFeature: SubscriptionsFeature,
     private val partnershipsHubUrlProvider: PartnershipsHubUrlProvider,
