+++ subscriptions/subscriptions-impl/src/main/res/values/donottranslate.xml
@@ -24,7 +24,7 @@
     <string name="planTierNamePro">Pro</string>
 
     <!-- Partnerships Hub -->
-    <string name="partnershipsSettingTitle">Partner Benefits</string>
+    <string name="partnerBenefitSettingTitle">Partner Benefits</string>
 
     <!-- Black Friday offer-->
     <string name="subscriptionSettingBlackFridayOffer">Save 40% on First Year</string>
